requirements.txt
djangorestframework 
drf-yasg 
djangorestframework-simplejwt
python-decouple
Pillow